move filtering functionality to separate package
