+++ exoscale/datasource_exoscale_compute_instance_list_test.go
@@ -1,7 +1,6 @@
 package exoscale
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"testing"
@@ -87,92 +86,6 @@ data "exoscale_compute_instance_list" "test" {
 			},
 		},
 	})
-}
-
-func TestComputeInstanceListFilterString(t *testing.T) {
-	attributeToMatch := "my-test-attr"
-	valueToMatch := "string-to-match"
-
-	dataToFilter := map[string]interface{}{
-		attributeToMatch: valueToMatch,
-	}
-
-	matchFn, err := createMatchStringFunc(valueToMatch)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	filter := createStringFilterFunc(attributeToMatch, matchFn)
-
-	if !checkForMatch(dataToFilter, []filterFunc{filter}) {
-		t.Error("should match")
-	}
-}
-
-func TestComputeInstanceListFilterRegex(t *testing.T) {
-	attributeToMatch := "my-test-attr"
-	valueToMatch := "string-123-to-match-by-regex"
-
-	dataToFilter := map[string]interface{}{
-		attributeToMatch: valueToMatch,
-	}
-
-	matchFn, err := createMatchStringFunc("/.*123.*/")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	filter := createStringFilterFunc(attributeToMatch, matchFn)
-
-	if !checkForMatch(dataToFilter, []filterFunc{filter}) {
-		t.Error("should match")
-	}
-}
-
-func TestComputeInstanceListFilterLabelsExactly(t *testing.T) {
-	labelToMatch := "my-label"
-
-	dataToFilter := map[string]interface{}{
-		"labels": map[string]string{
-			labelToMatch: "label-string-to-match",
-		},
-	}
-
-	labelsFilterProp := map[string]interface{}{
-		labelToMatch: "label-string-to-match",
-	}
-
-	filter, err := createMapStrToStrFilterFunc(context.Background(), labelsFilterProp)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !checkForMatch(dataToFilter, []filterFunc{filter}) {
-		t.Error("should match")
-	}
-}
-
-func TestComputeInstanceListFilterLabelsRegex(t *testing.T) {
-	labelToMatch := "my-label"
-
-	dataToFilter := map[string]interface{}{
-		"labels": map[string]string{
-			labelToMatch: "label-string-to-match",
-		},
-	}
-
-	labelsFilterProp := map[string]interface{}{
-		labelToMatch: "/.*-to.*-/",
-	}
-
-	filter, err := createMapStrToStrFilterFunc(context.Background(), labelsFilterProp)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !checkForMatch(dataToFilter, []filterFunc{filter}) {
-		t.Error("should match")
-	}
 }
 
 func testAccDataSourceComputeInstanceListAttributes(ds string, expected testAttrs) resource.TestCheckFunc {
